remove apt cache update step
This is a quick fix so that I can run this role as a non-root user. Everything can run as an unprivileged user except for updating the apt cache, and installing dependencies from apt. With this change when running as non-root the role will fail if dependencies are not present; otherwise everything will work.